% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fix.R
\name{create_fix_gko}
\alias{create_fix_gko}
\title{Generating group fixed-X knockoff variables}
\usage{
create_fix_gko(
  x,
  y = NULL,
  group,
  method = c("equi", "sdp"),
  sigma = NULL,
  randomize = FALSE
)
}
\arguments{
\item{x}{A \code{n}-by-\code{p} matrix of original variables}

\item{y}{The \code{n}-dimensional response vector. If the error standard error \code{sigma} is not specified, then \code{y} is required to estimate \code{sigma}.}

\item{group}{A \code{p}-dimensional vector of consecutive integers describing the group information of the variables. For example, group <- c(1, 1, 2, 2, 3, 3) means the first two vars belong to group 1, second two vars belong to group 2, and the last two vars belong to group 3}

\item{method}{Method of constructing fixed-X knockoffs.}

\item{sigma}{The error standard deviation. If specified as \code{NULL} (as default), then the program will estimate it. In such a case, the response vector \code{y} has to be supplied to the program.}

\item{randomize}{An indicator of whether the knockoffs are constructed deterministically (default) or randomized.}
}
\value{
An object of class "gknockoff.variables", which contains the following components:
\describe{
\item{\code{x}}{the \code{n}-by-\code{p} original design matrix. This could be transformed or augmented in the case where \code{n} < \code{2p}}
\item{\code{xk}}{the \code{n}-by-\code{p} constructed group fixed-X knockoff variables}
\item{\code{y}}{the \code{n}-dim response vector, if specified into the program. This could be the augmented if \code{n} < \code{2p}.}
}
}
\description{
This function samples multivariate Gaussian fixed-X knockoff variables
}
\examples{
set.seed(123)
n <- 200
p <- 10
gr <- c(rep(1, 5), rep(2, 2), 3, rep(4, 2))
x <- matrix(rnorm(n * p), n, p)
beta <- rep(0, p)
beta[gr == 2] <- 2
beta[gr == 3] <- 5
y <- x \%*\% beta + rnorm(n)
gko <- create_fix_gko(x = x, y = y, group = gr)
}
