% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_gko}
\alias{filter_gko}
\title{Generating group fixed-X knockoff variables}
\usage{
filter_gko(
  x,
  y,
  group,
  method = "equi",
  loss = "ls",
  q = 0.2,
  plus = TRUE,
  sigma = NULL,
  randomize = FALSE
)
}
\arguments{
\item{x}{A \code{n}-by-\code{p} matrix of original variables}

\item{y}{The \code{n}-dimensional response vector. If the error standard error \code{sigma} is not specified, then \code{y} is required to estimate \code{sigma}.}

\item{group}{A \code{p}-dimensional vector of consecutive integers describing the group information of the variables. For example, group <- c(1, 1, 2, 2, 3, 3) means the first two vars belong to group 1, second two vars belong to group 2, and the last two vars belong to group 3. The 0-th group are fixed and not to be selected.}

\item{method}{Method of constructing fixed-X knockoffs.}

\item{loss}{a character string specifying the loss function to use in fitting the group lasso problem. Default is \code{"ls"} for least squares loss.}

\item{q}{A pre-specified nominal level for the controled group FDR}

\item{plus}{Default to be \code{TRUE}, which gives the knockoff+ procedure that controls the FDR under \code{q}. If \code{plus = FALSE}, then the output does not technically controls FDR under \code{q}, but it controls the modeified group FDR under \code{q}.}

\item{sigma}{The error standard deviation. If specified as \code{NULL} (as default), then the program will estimate it. In such a case, the response vector \code{y} has to be supplied to the program.}

\item{randomize}{An indicator of whether the knockoffs are constructed deterministically (default) or randomized.}
}
\value{
An object of class "gknockoff.result", which contains the following components:
\describe{
\item{\code{call}}{function call}
\item{\code{x}}{the \code{n}-by-\code{p} original design matrix. This could be transformed or augmented in the case where \code{n} < \code{2p}}
\item{\code{xk}}{the \code{n}-by-\code{p} constructed group fixed-X knockoff variables}
\item{\code{y}}{the \code{n}-dim response vector, if specified into the program. This could be the augmented if \code{n} < \code{2p}}
\item{\code{group}}{grouping information}
\item{\code{stat}}{group knockoff statistics}
\item{\code{threshold}}{the computed thresholed for filter}
\item{\code{selected}}{the set of selected groups}
}
}
\description{
This function samples multivariate Gaussian fixed-X knockoff variables
}
\examples{
set.seed(123)
n <- 200
p <- 10
gr <- c(rep(0, 2), rep(1, 3), rep(2, 2), 3, rep(4, 2))
x <- matrix(rnorm(n * p), n, p)
beta <- rep(0, p)
beta[gr == 0] <- 10
beta[gr == 2] <- 2
beta[gr == 3] <- 5
y <- x \%*\% beta + rnorm(n)
selected <- filter_gko(x = x, y  = y, group = gr)
}
