% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_gglasso_lambda.R
\name{stat_gglasso_lambda}
\alias{stat_gglasso_lambda}
\title{Compute group lasso statistics for group knockoffs}
\usage{
stat_gglasso_lambda(xo = NULL, gko, group, loss = "ls")
}
\arguments{
\item{xo}{The offset variable matrix}

\item{gko}{An object from group-knockoff construction.}

\item{group}{A \code{p}-dimensional vector of consecutive integers describing the group information of the variables. For example, group <- c(1, 1, 2, 2, 3, 3) means the first two vars belong to group 1, second two vars belong to group 2, and the last two vars belong to group 3. The 0-th group are fixed and not to be selected.}

\item{loss}{a character string specifying the loss function to use in fitting the group lasso problem. Default is \code{"ls"} for least squares loss.}
}
\description{
This function computes the group knockoff statistics used for the filter step based on the maximum value of lambdas for which the group of variables have non-zero coefficients estimates.
}
\details{
This function computes and returns a vector \code{stat}
\deqn{stat_j = max(\lambda_j, \tau_j) * sgn(\lambda_j - \tau_j),}
where \eqn{\lambda_j} and \eqn{\tau_j} are the largest values of lambdas at which the j-th group of variables, and the corresponding group knockoff variables, enter the model (i.e., coefficients become nonzero), respectively.
}
\examples{
set.seed(123)
n <- 200
p <- 10
gr <- c(rep(1, 5), rep(2, 2), 3, rep(4, 2))
x <- matrix(rnorm(n * p), n, p)
beta <- rep(0, p)
beta[gr == 2] <- 2
beta[gr == 3] <- 5
y <- x \%*\% beta + rnorm(n)
gko <- create_fix_gko(x = x, y = y, group = gr)
stat <- stat_gglasso_lambda(gko = gko, group = gr)
}
