% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{thresh_gko}
\alias{thresh_gko}
\title{Compute filter threshold for group knockoffs}
\usage{
thresh_gko(stat, q, plus = TRUE)
}
\arguments{
\item{stat}{A \code{m} dimensional vector of group knockoff statistics}

\item{q}{A pre-specified nominal level for the controlled group FDR}

\item{plus}{Default to be \code{TRUE}, which gives the knockoff+ procedure that controls the FDR under \code{q}. If \code{plus = FALSE}, then the output does not technically controls FDR under \code{q}, but it controls the modified group FDR under \code{q}.}
}
\value{
a threshold
}
\description{
This function computes the threshold of the group knockoff statistics used for the filter step
}
\examples{
set.seed(123)
n <- 200
p <- 10
gr <- c(rep(1, 5), rep(2, 2), 3, rep(4, 2))
x <- matrix(rnorm(n * p), n, p)
beta <- rep(0, p)
beta[gr == 2] <- 2
beta[gr == 3] <- 5
y <- x \%*\% beta + rnorm(n)
gko <- create_fix_gko(x = x, y = y, group = gr)
stat <- stat_gglasso_lambda(x = gko$x, xk = gko$xk, y = gko$y, group = gr)
t <- thresh_gko(stat, q = 0.2)
}
